move logo size to _base.scss
If you want add another logo it could be useful to change the size of it
with a custom css overwrite. With the old version this isn't possible
because the size defined in the img tag is always more significant. This
commit removes the height and width from the img tag and moves it to the
_base.scss file.